Delay cpu copy till end of epoch
Asynchronous listeners are not as effective in improving training performance. The main thing
that matters is to avoid copy to cpu. That means logging anything at the end of a batch must
be avoid. The main thing that this PR tries to do is to delay it to the end of the epoch. This
also means we only report loss and other evaluators only at the end of an epoch. From what I
have seen so far, it does help speeding up training, although it is difficult to quantify as
it depends on the model being trained, the batch size, the number of GPUs.

Change-Id: I8b01082688ce575889d57abf8ede1796de770cb5